% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_ratio.r
\name{calcul_ratio}
\alias{calcul_ratio}
\title{Une fonction permettant de calculer des ratios par rapport à des sur-populations.}
\usage{
calcul_ratio(data, var, by, new_name = NULL, ...)
}
\description{
Une fonction permettant de calculer des ratios par rapport à des sur-populations.
}
\examples{
#Rien
}

