% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricature_clustering.r
\name{caricature_clustering}
\alias{caricature_clustering}
\title{clustering}
\usage{
caricature_clustering(data, var_to_caricature, var_X, weight = NULL,
  nb_cluster = 5, min_pop = round(min(NROW(data)/1000,
  NROW(data)/nb_cluster/10)), names_bloc = NULL, ...)
}
\arguments{
\item{data}{est une \code{data.frame}.}

\item{var_to_caricature}{est un vecteur de \code{character} de noms de colonnes \code{data} constituant la ou les variables à caricaturer (dans le cas de plusieurs variables celles-ci seront sytnthéisée en une typologie  de \code{nb_cluster classes} au prélaable).}

\item{var_X}{est un vecteur de \code{character} de noms de colonnes \code{data} constituant la ou les variables explicatives pour la caricature.}

\item{weight}{soit \code{NULL} (pas  de pondération des indivudus) soit un vecteur de \code{numeric} positif de même longueur que \code{nrow(data)} .}

\item{nb_cluster}{est un \code{integer} représentant le nombre de clusters à construire.}

\item{min_pop}{est soit \code{NULL} (pas de contrainte) soit un \code{integer} indiquant la taille minimale de chaque cluster.}

\item{...}{sont des paramètres complémentaires non utilisés.}
}
\description{
Une fonction permettant de réaliser toutes les analyses factorielles.
Version 0.01 de la fonction Clustering, cette version est softe et ne comprend pas en compte les contraintes sur la taille des clusters
}
\examples{
#Pas d'exemple pour le moment
}

