% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{clustering}
\alias{clustering}
\title{clustering}
\usage{
clustering(data, var = colnames(data), weight = NULL, nb_cluster = 5,
  min_pop = round(min(NROW(data)/1000, NROW(data)/nb_cluster/10)),
  max_pop = round(2 * NROW(data)/nb_cluster), seed = 123, ...)
}
\arguments{
\item{data}{est une \code{data.frame}.}

\item{var}{est est un vecteur de \code{character} de noms de colonnes \code{data} constituant les variables à prendre en compte dans le clustering. Les variables sont ici supposées être toutes nummériques.}

\item{weight}{soit \code{NULL} (pas  de pondération des indivudus) soit un vecteur de \code{numeric} positif de même longueur que \code{nrow(data)} .}

\item{nb_cluster}{est un \code{integer} représentant le nombre de clusters à construire.}

\item{min_pop}{est soit \code{NULL} (pas de contrainte) soit un \code{integer} indiquant la taille minimale de chaque cluster.}

\item{max_pop}{est soit \code{NULL} (pas de contrainte) soit un \code{integer} indiquant la taille maximale de chaque cluster.}

\item{...}{sont des paramètres complémentaires dont la nature dépend de la méthode (à définir donc selon les besoin des développements).}
}
\description{
Une fonction permettant de réaliser toutes les analyses factorielles.
Version 0.01 de la fonction Clustering, cette version est softe et ne comprend pas en compte les contraintes sur la taille des clusters
}
\examples{
#Pas d'exemple pour le moment
}

