% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_extreme.r
\name{DoubleMAD}
\alias{DoubleMAD}
\title{Une fonction permettant de cacluler des détecter et traiter les valeurs extrêmes dans une table}
\usage{
DoubleMAD(x, zero.mad.action = "warn")
}
\description{
Une fonction permettant de cacluler des détecter et traiter les valeurs extrêmes dans une table
}
\examples{
#Rien
}

