% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.r
\name{imputation}
\alias{imputation}
\title{Une fonction permettant d'imputer les valeurs manquantes d'une table}
\usage{
imputation(data, var = colnames(data), var_comp = colnames(data),
  method = c("Médiane / Mode", "Hot Deck", "MCMC", "Forêts Aléatoires"),
  seuil = NULL, ...)
}
\description{
Une fonction permettant d'imputer les valeurs manquantes d'une table
}
\examples{
#Rien
}

