% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profil_variable.r
\name{profil_variable}
\alias{profil_variable}
\title{Profil d'une variable}
\usage{
profil_variable(data, var, ...)
}
\arguments{
\item{data}{est une \code{data.frame}.}

\item{var}{est un vecteur de \code{character} de noms de colonnes présents dans `data` et dont les types sont tous \code{numeric} ou \code{factor}.}

\item{...}{sont des paramètres complémentaires dont la nature dépend de la méthode (à définir donc selon les besoin des développements).}
}
\description{
Fonction permettant de créer des résumés sous forme de texte (affichables) des distributions des variables d'une table.
}
\examples{
#Pas d'exemple pour le moment
}

