% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport1.r
\name{rapport1}
\alias{rapport1}
\title{Profil d'une variable}
\usage{
rapport1(file, out, data_role, lang = "fr", ...)
}
\arguments{
\item{out}{est le résultat fournit par un appel de la fonction \code{factor_analysis}.}

\item{data_role}{est une \code{data.frame} issue directement de l'interface et fournissant des informaitons sur les stauts des variabkes.}

\item{...}{sont des paramètres complémentaires dont la nature dépend de la méthode (à définir donc selon les besoin des développements).}
}
\description{
Fonction permettant de créer des résumés sous forme de texte (affichables) des distributions des variables d'une table.
}
\examples{
#Pas d'exemple pour le moment
}

