% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport2.r
\name{rapport2}
\alias{rapport2}
\title{Profil d'une variable}
\usage{
rapport2(file, detail, param_init, data_role, ttl, id_analysis, init, all_data,
  nom, weight = NULL, var_selected_parangon = NULL, lang = "fr", ...)
}
\arguments{
\item{data_role}{est une \code{data.frame} issue directement de l'interface et fournissant des informaitons sur les stauts des variabkes.}

\item{...}{sont des paramètres complémentaires dont la nature dépend de la méthode (à définir donc selon les besoin des développements).}

\item{out}{est le résultat fournit par un appel de la fonction \code{factor_analysis}.}
}
\description{
Fonction permettant de créer des résumés sous forme de texte (affichables) des distributions des variables d'une table.
}
\examples{
#Pas d'exemple pour le moment
}

