% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_clustering.r
\name{split_clustering}
\alias{split_clustering}
\title{Une fonction permettant de réaliser toutes les analyses factorielles.}
\usage{
split_clustering(data, var = colnames(data), weight = NULL, nb_decoup = 2,
  min_pop = round(min(NROW(data)/1000, NROW(data)/nb_decoup/10)), ...)
}
\arguments{
\item{data}{est une\code{data.frame}.}

\item{var}{est un vecteur de \code{character} de noms de colonnes `data` constituant les variables actives de l'analyses./}

\item{weight}{soit \code{NULL} (pas  de pondération des indivudus) soit un vecteur de \code{numeric} positif de même longueur que \code{nrow(data)} .}

\item{nb_decoup}{est un \code{integer} représentant le nombre de clusters à construire.}

\item{min_pop}{est soit \code{NULL} (pas de contrainte) soit un \code{integer} indiquant la taille minimale de chaque cluster.}

\item{...}{sont des paramètres complémentaires dont la nature dépend de la méthode (à définir donc selon les besoin des développements).}
}
\description{
Une fonction permettant de réaliser toutes les analyses factorielles.
}
\examples{
#Pas d'exemple pour le moment
}

